Replace snprintf with string and ostringstream
These uses are probably safe from a buffer overflow perspective but
can cause data race issues in logging due to static buffers.
