+++ src/s3fs_logger.h
@@ -32,12 +32,6 @@
 #define S3FS_CLOCK_MONOTONIC    CLOCK_MONOTONIC
 #endif
 
-#if defined(__APPLE__)
-#define S3FSLOG_TIME_FMT        "%s.%03dZ "
-#else
-#define S3FSLOG_TIME_FMT        "%s.%03ldZ "
-#endif
-
 //-------------------------------------------------------------------
 // S3fsLog class
 //-------------------------------------------------------------------
@@ -62,7 +56,6 @@ class S3fsLog
         static s3fs_log_level debug_level;
         static FILE*          logfp;
         static std::string*   plogfile;
-        static char           current_time[64];
         static bool           time_stamp;
 
     protected:
@@ -87,26 +80,7 @@ class S3fsLog
                      LEVEL_ERR  == (level & LEVEL_DBG) ? LOG_ERR     : LOG_CRIT );
         }
 
-        static const char* GetCurrentTime()
-        {
-            if(time_stamp){
-                struct timeval  now;
-                struct timespec tsnow;
-                struct tm res;
-                char   tmp[32];
-                if(-1 == clock_gettime(S3FS_CLOCK_MONOTONIC, &tsnow)){
-                    now.tv_sec  = tsnow.tv_sec;
-                    now.tv_usec = (tsnow.tv_nsec / 1000);
-                }else{
-                    gettimeofday(&now, NULL);
-                }
-                strftime(tmp, sizeof(tmp), "%Y-%m-%dT%H:%M:%S", gmtime_r(&now.tv_sec, &res));
-                snprintf(current_time, sizeof(current_time), S3FSLOG_TIME_FMT, tmp, (now.tv_usec / 1000));
-            }else{
-                current_time[0] = '\0';
-            }
-            return current_time;
-        }
+        static std::string GetCurrentTime();
 
         static const char* GetLevelString(s3fs_log_level level)
         {
@@ -172,7 +146,7 @@ class S3fsLog
             if(S3fsLog::IsS3fsLogLevel(level)){ \
                 if(foreground || S3fsLog::IsSetLogFile()){ \
                     S3fsLog::SeekEnd(); \
-                    fprintf(S3fsLog::GetOutputLogFile(), "%s%s%s:%s(%d): " fmt "%s\n", S3fsLog::GetCurrentTime(), S3fsLog::GetLevelString(level), __FILE__, __func__, __LINE__, __VA_ARGS__); \
+                    fprintf(S3fsLog::GetOutputLogFile(), "%s%s%s:%s(%d): " fmt "%s\n", S3fsLog::GetCurrentTime().c_str(), S3fsLog::GetLevelString(level), __FILE__, __func__, __LINE__, __VA_ARGS__); \
                     S3fsLog::Flush(); \
                 }else{ \
                     syslog(S3fsLog::GetSyslogLevel(level), "%s%s:%s(%d): " fmt "%s", instance_name.c_str(), __FILE__, __func__, __LINE__, __VA_ARGS__); \
@@ -185,7 +159,7 @@ class S3fsLog
             if(S3fsLog::IsS3fsLogLevel(level)){ \
                 if(foreground || S3fsLog::IsSetLogFile()){ \
                     S3fsLog::SeekEnd(); \
-                    fprintf(S3fsLog::GetOutputLogFile(), "%s%s%s%s:%s(%d): " fmt "%s\n", S3fsLog::GetCurrentTime(), S3fsLog::GetLevelString(level), S3fsLog::GetS3fsLogNest(nest), __FILE__, __func__, __LINE__, __VA_ARGS__); \
+                    fprintf(S3fsLog::GetOutputLogFile(), "%s%s%s%s:%s(%d): " fmt "%s\n", S3fsLog::GetCurrentTime().c_str(), S3fsLog::GetLevelString(level), S3fsLog::GetS3fsLogNest(nest), __FILE__, __func__, __LINE__, __VA_ARGS__); \
                     S3fsLog::Flush(); \
                 }else{ \
                     syslog(S3fsLog::GetSyslogLevel(level), "%s%s" fmt "%s", instance_name.c_str(), S3fsLog::GetS3fsLogNest(nest), __VA_ARGS__); \
@@ -197,7 +171,7 @@ class S3fsLog
         do{ \
             if(foreground || S3fsLog::IsSetLogFile()){ \
                 S3fsLog::SeekEnd(); \
-                fprintf(S3fsLog::GetOutputLogFile(), "%s[CURL DBG] " fmt "%s\n", S3fsLog::GetCurrentTime(), __VA_ARGS__); \
+                fprintf(S3fsLog::GetOutputLogFile(), "%s[CURL DBG] " fmt "%s\n", S3fsLog::GetCurrentTime().c_str(), __VA_ARGS__); \
                 S3fsLog::Flush(); \
             }else{ \
                 syslog(S3fsLog::GetSyslogLevel(S3fsLog::LEVEL_CRIT), "%s" fmt "%s", instance_name.c_str(), __VA_ARGS__); \
@@ -221,7 +195,7 @@ class S3fsLog
         do{ \
             if(foreground || S3fsLog::IsSetLogFile()){ \
                 S3fsLog::SeekEnd(); \
-                fprintf(S3fsLog::GetOutputLogFile(), "%s%s%s%s:%s(%d): " fmt "%s\n", S3fsLog::GetCurrentTime(), S3fsLog::GetLevelString(S3fsLog::LEVEL_INFO), S3fsLog::GetS3fsLogNest(0), __FILE__, __func__, __LINE__, __VA_ARGS__, ""); \
+                fprintf(S3fsLog::GetOutputLogFile(), "%s%s%s%s:%s(%d): " fmt "%s\n", S3fsLog::GetCurrentTime().c_str(), S3fsLog::GetLevelString(S3fsLog::LEVEL_INFO), S3fsLog::GetS3fsLogNest(0), __FILE__, __func__, __LINE__, __VA_ARGS__, ""); \
                 S3fsLog::Flush(); \
             }else{ \
                 syslog(S3fsLog::GetSyslogLevel(S3fsLog::LEVEL_INFO), "%s%s" fmt "%s", instance_name.c_str(), S3fsLog::GetS3fsLogNest(0), __VA_ARGS__, ""); \
@@ -232,7 +206,7 @@ class S3fsLog
         do{ \
             if(foreground || S3fsLog::IsSetLogFile()){ \
                 S3fsLog::SeekEnd(); \
-                fprintf(S3fsLog::GetOutputLogFile(), "%s%s" fmt "%s\n", S3fsLog::GetCurrentTime(), S3fsLog::GetLevelString(S3fsLog::LEVEL_INFO), __VA_ARGS__, ""); \
+                fprintf(S3fsLog::GetOutputLogFile(), "%s%s" fmt "%s\n", S3fsLog::GetCurrentTime().c_str(), S3fsLog::GetLevelString(S3fsLog::LEVEL_INFO), __VA_ARGS__, ""); \
                 S3fsLog::Flush(); \
             }else{ \
                 syslog(S3fsLog::GetSyslogLevel(S3fsLog::LEVEL_INFO), "%s" fmt "%s", instance_name.c_str(), __VA_ARGS__, ""); \
